Linux: Changed the type of "handle" from "int" to "unsigned long"

+++ Base/PLCore/include/PLCore/PLCoreLinux.h
@@ -182,7 +182,7 @@ namespace PLCore {
 
 	typedef size_t				uint_ptr;
 
-	typedef int					handle;
+	typedef unsigned long		handle;
 
 
 }
